hardware/sensor_interface.py
import yaml
import serial
import time
from pathlib import Path
import logging

logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

class SensorInterface:
    def __init__(self, config_path):
        with open(config_path, 'r') as f:
            self.config = yaml.safe_load(f)
        
        self.nanopore_port = self.config['nanopore_device']['port']
        self.baud_rate = self.config['nanopore_device']['baud_rate']
        self.serial_connection = None
    
    def connect_to_device(self):
        """Establish connection to the hardware device"""
        try:
            self.serial_connection = serial.Serial(
                port=self.nanopore_port,
                baudrate=self.baud_rate,
                timeout=1
            )
            logger.info(f"Connected to device on port {self.nanopore_port}")
            return True
        except serial.SerialException as e:
            logger.error(f"Failed to connect to device: {e}")
            return False
    
    def disconnect(self):
        """Safely disconnect from the hardware"""
        if self.serial_connection and self.serial_connection.is_open:
            self.serial_connection.close()
            logger.info("Disconnected from device")
    
    def read_sensor_data(self):
        """Read data from physical sensors (to be implemented)"""
        if not self.serial_connection or not self.serial_connection.is_open:
            logger.error("No connection to device")
            return None
        
        try:
            # Placeholder for actual sensor reading implementation
            # This will be replaced with actual sensor communication protocol
            raw_data = self.serial_connection.readline()
            return self._parse_sensor_data(raw_data)
        except Exception as e:
            logger.error(f"Error reading sensor data: {e}")
            return None
    
    def _parse_sensor_data(self, raw_data):
        """Parse raw sensor data into structured format (to be implemented)"""
        # Placeholder for data parsing logic
        # Will be implemented based on actual sensor protocol
        try:
            # Example parsing (to be replaced with actual protocol)
            decoded = raw_data.decode('utf-8').strip()
            if not decoded:
                return None
            
            # Assuming CSV format: timestamp,co2,ph,moisture
            values = decoded.split(',')
            if len(values) != 4:
                return None
            
            return {
                'timestamp': float(values[0]),
                'co2_ppm': float(values[1]),
                'ph': float(values[2]),
                'moisture_percent': float(values[3])
            }
        except Exception as e:
            logger.error(f"Error parsing sensor data: {e}")
            return None
    
    def configure_device(self, settings):
        """Configure device settings (to be implemented)"""
        # Placeholder for device configuration
        # Will be implemented based on actual hardware specifications
        pass

if __name__ == "__main__":
    config_path = "../configs/simulation_config.yaml"
    interface = SensorInterface(config_path)
    
    if interface.connect_to_device():
        try:
            # Example usage
            data = interface.read_sensor_data()
            if data:
                print("Sensor readings:", data)
        finally:
            interface.disconnect()