requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.2
torch>=1.9.0
pyyaml>=5.4.1
jupyter>=1.0.0
matplotlib>=3.4.0
ipykernel>=6.0.0
h5py>=3.7.0
scipy>=1.7.0
pyserial>=3.5
tqdm>=4.65.0
seaborn>=0.11.0